add VideoMetadata Module
