--- libraries/videometadata/src/main/java/me/pitok/videometadata/di/VideoMetadataComponent.kt
@@ -1,18 +1,0 @@
-package me.pitok.videometadata.di
-
-import android.media.MediaMetadataRetriever
-import coil.ImageLoader
-import dagger.Component
-import me.pitok.androidcore.components.AndroidCoreComponent
-import me.pitok.dependencyinjection.library.LibraryScope
-
-@LibraryScope
-@Component(
-    modules = [VideoMetadataModule::class],
-    dependencies = [AndroidCoreComponent::class]
-)
-interface VideoMetadataComponent {
-    fun exposeMMDR(): MediaMetadataRetriever
-
-    fun exposeCoilVideoImageLoader(): ImageLoader
-}
