+++ libraries/videometadata/src/main/java/me/pitok/videometadata/di/builder/VideoMetadataComponentBuilder.kt
@@ -1,7 +1,10 @@
-package me.pitok.videometadata.di
+package me.pitok.videometadata.di.builder
 
 import me.pitok.androidcore.builder.AndroidCoreComponentBuilder
 import me.pitok.dependencyinjection.ComponentBuilder
+import me.pitok.videometadata.di.components.DaggerVideoMetadataComponent
+import me.pitok.videometadata.di.components.VideoMetadataComponent
+import me.pitok.videometadata.di.modules.VideoMetadataModule
 
 object VideoMetadataComponentBuilder: ComponentBuilder<VideoMetadataComponent>(){
     override fun initComponent(): VideoMetadataComponent {
@@ -10,5 +13,4 @@ object VideoMetadataComponentBuilder: ComponentBuilder<VideoMetadataComponent>()
             .videoMetadataModule(VideoMetadataModule())
             .build()
     }
-
 }
