+++ libraries/videometadata/src/main/java/me/pitok/videometadata/requests/FolderVideosRequest.kt
@@ -1,4 +1,4 @@
-package me.pitok.videolist.requests
+package me.pitok.videometadata.requests
 
 import android.content.ContentResolver
